move skl AFE in the pipeline file (since it is not a sklearn op)
