+++ doc/html/hummingbird/ml/operator_converters/sklearn/sv.html
@@ -303,18 +303,10 @@ register_converter(&#34;SklearnNuSVC&#34;, convert_sklearn_svc_model)</code></pr
 <h3>Methods</h3>
 <dl>
 <dt id="hummingbird.ml.operator_converters.sklearn.sv.SVC.forward"><code class="name flex">
-<span>def <span class="ident">forward</span></span>(<span>self, x)</span>
+<span>def <span class="ident">forward</span></span>(<span>self, x) -> Callable[..., Any]</span>
 </code></dt>
 <dd>
-<div class="desc"><p>Defines the computation performed at every call.</p>
-<p>Should be overridden by all subclasses.</p>
-<div class="admonition note">
-<p class="admonition-title">Note</p>
-<p>Although the recipe for forward pass needs to be defined within
-this function, one should call the :class:<code>Module</code> instance afterwards
-instead of this since the former takes care of running the
-registered hooks while the latter silently ignores them.</p>
-</div></div>
+<div class="desc"></div>
 <details class="source">
 <summary>
 <span>Expand source code</span>
