+++ doc/html/hummingbird/ml/supported.html
@@ -157,14 +157,7 @@ def _build_sklearn_operator_list():
         from sklearn.svm import LinearSVC, SVC, NuSVC
 
         # Preprocessing
-        from sklearn.preprocessing import (
-            MaxAbsScaler,
-            MinMaxScaler,
-            Normalizer,
-            OneHotEncoder,
-            RobustScaler,
-            StandardScaler,
-        )
+        from sklearn.preprocessing import MaxAbsScaler, MinMaxScaler, Normalizer, OneHotEncoder, RobustScaler, StandardScaler
 
         # Features
         from sklearn.feature_selection import SelectKBest, VarianceThreshold
