+++ hummingbird/ml/operator_converters/sklearn/pipeline.py
@@ -13,6 +13,7 @@ from onnxconverter_common.registration import register_converter
 import torch
 
 from .. import constants
+from .._array_feature_extractor_implementations import ArrayFeatureExtractor
 from .._base_operator import BaseOperator
 
 
@@ -32,6 +33,24 @@ class Multiply(BaseOperator, torch.nn.Module):
 
     def forward(self, x):
         return x * self.score
+
+
+def convert_sklearn_array_feature_extractor(operator, device, extra_config):
+    """
+    Converter for `sklearn.feature_selection.VarianceThreshold`.
+
+    Args:
+        operator: An operator wrapping a `sklearn.feature_selection.VarianceThreshold` model
+        device: String defining the type of device the converted operator should be run on
+        extra_config: Extra configuration used to select the best conversion strategy
+
+    Returns:
+        A PyTorch model
+    """
+    assert operator is not None
+
+    indices = operator.column_indices
+    return ArrayFeatureExtractor(np.ascontiguousarray(indices), device)
 
 
 def convert_sklearn_concat(operator, device=None, extra_config={}):
@@ -70,5 +89,6 @@ def convert_sklearn_multiply(operator, device=None, extra_config={}):
     return Multiply(score)
 
 
+register_converter("SklearnArrayFeatureExtractor", convert_sklearn_array_feature_extractor)
 register_converter("SklearnConcat", convert_sklearn_concat)
 register_converter("SklearnMultiply", convert_sklearn_multiply)
